Update about section
Move navbar to Header section and edit About page.
