+++ src/styles/About.css
@@ -1,3 +1,24 @@
 .aboutCon {
     background-color: #D11149;
+    min-height: 50vh;
+}
+
+.aboutRow {
+    padding: 20px 0 0 0;
+    height: 100%;
+}
+
+.aboutCards {
+    height: 100%;
+}
+
+.myPic {
+    width: 500px;
+    height: 500px;
+    border: solid 5px black;
+    margin: 20px 0 50px 0;
+}
+
+.profile {
+    background-color: #1A8FE3;
 }
